Improve organization deletion in user module
Added a condition to check if there are any organizations related to the user before proceeding with deletion. This would prevent unnecessary calls to deleteOrganizations function when there are no organizations associated with the user, thereby enhancing the efficiency of the user deletion process.

+++ src/lib/models/user.ts
@@ -173,12 +173,14 @@ export async function deleteUser(id: number): Promise<void> {
 	// Get all organizations related to this user, along with their number of owners.
 	const organizationsToDelete = await getUsersDependantOrganizations(id);
 
-	// Filter to only organizations which have a single owner (this user), and map to their ids.
-	const organizationsToDeleteIds = organizationsToDelete
-		.filter(({_count: {owners}}) => owners === 1)
-		.map(({id}) => id);
+	if (organizationsToDelete.length > 0) {
+		// Filter to only organizations which have a single owner (this user), and map to their ids.
+		const organizationsToDeleteIds = organizationsToDelete
+			.filter(({_count: {owners}}) => owners === 1)
+			.map(({id}) => id);
 
-	await deleteOrganizations(organizationsToDeleteIds);
+		await deleteOrganizations(organizationsToDeleteIds);
+	}
 
 	await prisma.$transaction(
 		[
